Explicitly add 'libclass-c3-xs-perl' as a Depends in debian/control.
According to discussions with upstream; although libclass-c3-xs-perl is only a
Recommends of the Moose module, it provides a '10 fold' increase in performance
of the LedgerSMB application when it is installed. So even in the case where
installations are being done without Recommends, it should still be available.
Explicitly add it as a Depends to ensure that it will be.