Remove the 'override_dh_auto_test' target from debian/rules.
Have had an empty 'override_dh_auto_test' target in debian/rules since the
packaging of LedgerSMB was started, so that the testing would not interfere
with package building.  But it's long past time to see how that goes, and
since there is 'Makefile' and it does have a 'test' target, should only need
to remove that override in order for the tests to run.